backend/requirements.txt
# inFlow Shield API — requirements
# Install: pip install -r requirements.txt

fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1

# inflow_shield_lib (replaces llm_guard)
inflow-shield-lib @ git+https://github.com/cyberhunter0304/inflow-shield-lib.git

# PII detection
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
spacy==3.7.5

# ML models (required by inflow_shield_lib PromptInjection + Toxicity)
transformers==4.46.3
torch==2.5.1
